utils/find-playlist-videos-not-archived/pyproject.toml
[project]
name = "find-playlist-videos-not-archived"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "AcipenserSturio", email = "AcipenserSturio@users.noreply.github.com"},
]
dependencies = ["pyyaml>=6.0.2", "yt-dlp>=2025.5.22", "pandas>=2.2.3"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
